Fixed compiler warning and podspec

+++ BaasBox-iOS-SDK/BAAUser.m
@@ -206,7 +206,7 @@
     
     NSArray *exclude = @[@"authenticationToken", @"pushNotificationToken", @"pushEnabled", @"roles"];
     NSMutableDictionary *result = [NSMutableDictionary dictionary];
-    NSUInteger propertiesCount;
+    unsigned int propertiesCount;
     objc_property_t *propertyList = class_copyPropertyList([self class], &propertiesCount);
     
     for (int i = 0 ; i < propertiesCount; i++) {
